refactor: check uses of Button in manufacturing package
