+++ packages/apps/manufacturing/src/components/organisms/index.js
@@ -22,7 +22,7 @@ export {default as LinkedManufacturingOrderCard} from './LinkedManufacturingOrde
 export {default as ManufacturingOrderCard} from './ManufacturingOrderCard/ManufacturingOrderCard';
 export {default as ManufacturingOrderHalfLabelCardList} from './ManufacturingOrderHalfLabelCardList/ManufacturingOrderHalfLabelCardList';
 export {default as ManufacturingOrderHeader} from './ManufacturingOrderHeader/ManufacturingOrderHeader';
-export {default as ManufacturingOrderIconButtonList} from './ManufacturingOrderIconButtonList/ManufacturingOrderIconButtonList';
+export {default as ManufacturingOrderButtons} from './ManufacturingOrderButtons/ManufacturingOrderButtons';
 export {default as ManufacturingOrderNotesCardList} from './ManufacturingOrderNotesCardList/ManufacturingOrderNotesCardList';
 export {default as ManufacturingOrderProductionOrderSetView} from './ManufacturingOrderProductionOrderSetView/ManufacturingOrderProductionOrderSetView';
 export {default as ManufacturingOrderSaleOrderSetView} from './ManufacturingOrderSaleOrderSetView/ManufacturingOrderSaleOrderSetView';
